% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{get_example_paper}
\alias{get_example_paper}
\title{Get path to example paper}
\usage{
get_example_paper(example = "example_paper.pdf")
}
\arguments{
\item{example}{Character string. Name of example file (default: "example_paper.pdf")}
}
\value{
Path to example file if it exists, otherwise downloads it from GitHub
}
\description{
Returns the path to the example paper included in the package source
(available on GitHub but not in CRAN builds).
}
\examples{
\dontrun{
paper_path <- get_example_paper()
doc <- pdf2txt_auto(paper_path, n_columns = 2)
}

}
