% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2txtAI.R
\name{remove_all_tables}
\alias{remove_all_tables}
\title{Remove All Types of Tables (Markdown and Plain Text)}
\usage{
remove_all_tables(text)
}
\arguments{
\item{text}{Character string containing text with tables}
}
\value{
Character string with all tables and table captions removed
}
\description{
Remove All Types of Tables (Markdown and Plain Text)
}
