% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2txtAI.R
\name{remove_figure_caps}
\alias{remove_figure_caps}
\title{Remove Figure Captions}
\usage{
remove_figure_caps(text)
}
\arguments{
\item{text}{Character string containing markdown text}
}
\value{
Character string with figure captions removed
}
\description{
Remove Figure Captions
}
